Replacement of content inside Home component
